Added new test to PHPunit testsuite changed linefeed detection
